rework filtering frame validation
Now, does not pre-validate the filtering frame: simply try to transform
to the filtering frame if it is provided, and continue without using
filtering frame if it is not available.

Helps in the case hri_face_detect starts before the robot's TF is fully
published -> the filtering will start in the correct frame as soon as it
become available.