Supporting depth images for face pose estimation
hri_face_detect now supports depth images to improve face
pose estimation. Now, if the use_depth parameter is set to true,
the node subscribes to the depth image and iterate over
the face keypoints to extract the face position. This improves
significantly the precision of the face pose estimation,
especially when the face is relatively distant from the camera
(> 1.5 meters). At this stage, the node assumes the depth image
to be aligned with the rgb image. This might change in the future.
The node supports both compressed (compressedDepth transport) and
raw depth images.